refactor [#436] 폴더 경로 수정 및 역할에 맞게 클래스 분리
