+++ src/test/java/org/sopt/confeti/restdocs/base/ElasticsearchTestConfiguration.java
@@ -1,4 +1,4 @@
-package org.sopt.confeti.domain.setlist.application;
+package org.sopt.confeti.restdocs.base;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
@@ -9,19 +9,19 @@ import org.springframework.data.elasticsearch.repository.config.EnableElasticsea
 
 @TestConfiguration
 @EnableElasticsearchRepositories(basePackages = "org.sopt.confeti.domain.elastic_search.infra")
-public class TestElasticsearchConfiguration extends ElasticsearchConfiguration {
+public class ElasticsearchTestConfiguration extends ElasticsearchConfiguration {
 
     @Bean
     @Primary
     @Override
     public ClientConfiguration clientConfiguration() {
         // 컨테이너가 실행 중인지 확인
-        if (!BaseControllerTest.elasticsearchContainer.isRunning()) {
+        if (!APIBaseTest.elasticsearchContainer.isRunning()) {
             throw new IllegalStateException("Elasticsearch container is not running!");
         }
 
         // BaseControllerTest의 elasticsearchContainer 참조
-        String httpHostAddress = BaseControllerTest.elasticsearchContainer.getHttpHostAddress();
+        String httpHostAddress = APIBaseTest.elasticsearchContainer.getHttpHostAddress();
 
         return ClientConfiguration.builder()
                 .connectedTo(httpHostAddress)
