+++ src/test/java/org/sopt/confeti/restdocs/base/TestDataManager.java
@@ -1,4 +1,4 @@
-package org.sopt.confeti.domain.setlist.application;
+package org.sopt.confeti.restdocs.base;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
